Fix ERB syntax for development server connection
When uncommenting the line (at the first hash), one would expect `connect_to_development_server` to reflect the result of `Rails.env.development?`, unless catching that the intended echo statement later in the same line was _also_ commented out. For the expected behavior, the `<%#` needs to also be uncommented as `<%=`, which is easy to miss and difficult to debug.

Fixes #54.